improved ef core configuration for entities

+++ ProjectF.Application/Auth/AuthUserCrudHandler.cs
@@ -110,7 +110,7 @@ namespace ProjectF.Application.Auth
                 _ => Error.New("couldn't find to country")
             };
 
-        public async Task<bool> UpdateUserCompany(long companyId, string userId)
+        public async Task<bool> UpdateUserCompany(int companyId, string userId)
         {
             var authUser = await _userManager.FindByIdAsync(userId);
             authUser.EditUser(authUser.Firstname, authUser.Lastname, companyId);
